requirements.txt
streamlit
ultralytics
torch
opencv-python-headless
Pillow